+++ CHANGES.rst
@@ -12,6 +12,7 @@ CHANGE HISTORY
 
 - Added set_scenes() and draw_next_frame() to allow asynchronous frameworks to
   use Screen.
+- Added Plasma renderer and sample code to use it.
 - Added background colour support to ColourImageFile.
 - Added support for multi-colour rendering using ${c,a,b} syntax.
 - Added highlight() method to Screen and Canvas.